+++ hexagonal/src/main/java/com/example/hexagonal/adapters/out/repository/entity/CustomerEntity.java
@@ -1,0 +1,16 @@
+package com.example.hexagonal.adapters.out.repository.entity;
+
+import lombok.Data;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Data
+@Document(collection = "customers")
+public class CustomerEntity {
+    
+    @Id
+    private String id;
+    private String name;
+    private AddressEntity address;
+    private String cpf;
+    private Boolean isValidCpf;
+}