+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Creating UseCase
 - Creating Client
 - Creating Adapter for address
+- Creating Repository